Remedies some issues
For the TrainController, I've added the actual meat to the getters and setters he needs.
For the TrackModel, changed the way my getPosition and setPosition as well as getBlock and setBlock are operating. Same with Coordinates
Need to figure out how to get the grade given the integer block ID for the current block.